requirements.txt
# CSC IRC Platform - Pinned Dependencies
# Generated 2026-04-11
# Install: pip install -r requirements.txt
# Then: pip install -e packages/csc-root -e packages/csc-log ... (see INSTALL_ORDER)

# ============================================================
# Third-party dependencies (pinned to tested versions)
# ============================================================

# Core
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
attrs==26.1.0
typing_extensions==4.15.0

# Crypto
cryptography==46.0.4
cffi==2.0.0
pycparser==3.0

# HTTP / networking
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
anyio==4.13.0
sniffio==1.3.1
certifi==2026.1.4
idna==3.11
urllib3==2.6.3
requests==2.32.5
charset-normalizer==3.4.4

# AI providers
anthropic==0.91.0
openai==2.31.0
google-generativeai==0.3.2
google-genai==1.72.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.2
google-auth==2.49.2
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
proto-plus==1.27.0
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
rsa==4.9.1

# AI support
jiter==0.13.0
distro==1.9.0
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
typing-inspection==0.4.2
tqdm==4.67.1
tenacity==9.1.4

# Misc
colorama==0.4.6
python-dotenv==1.2.2
toml==0.10.2
websockets==16.0

# VFS (git dependency)
enc-ext-vfs @ git+https://github.com/daveylongshaft/enc-ext-vfs.git

# ============================================================
# Dev / test (not required for runtime)
# ============================================================
# pytest==9.0.2
# ruff==0.15.9
